llm-simple-rag-chat/requirements.txt
python-dotenv
pandas
httpx

langchain_community
langchain_chroma
langchain_huggingface
langchain_google_genai
langchain_openai

google.generativeai # Required by langchain_google_genai
openai # Required by langchain_openai
tiktoken
mlflow
rank_bm25
unstructured
pymupdf
faiss-cpu
numpy<2 # Required by faiss-cpu
torch
